Issue #377 - use the VMUSB drivers in place rather than as copies
Import support for SIS3820 module from vmusb directly
