+++ main/usb/vmusb/daqconfig/C3820.h
@@ -16,15 +16,10 @@
 #ifndef __C3820_H
 #define __C3820_H
 
-#ifndef __CREADOUTHARDWARE_H
-#include "CReadoutHardware.h"
-#endif
-
-#ifndef __CRT_STDINT_H
+#include <CReadoutHardware.h>
 #include <stdint.h>
-#ifndef __CRT_STDINT_H
-#define __CRT_STDINT_H
-#endif
+#ifdef MVLC_GENERATOR
+#include <DeviceCommand.h>
 #endif
 
 // Forward class definitions:
@@ -32,7 +27,11 @@
 class CReadoutModule;
 class CVMUSB;
 class CVMUSBReadoutList;
-
+#ifdef MVLC_CONFIG
+namespace XXUSB {
+  class CConfigurableObject;
+}
+#endif
 /*!
    The C3820 is a scaler module that will be read out during scaler events.
    We will run the module in 32 bit latch mode.  In our case, the latch
@@ -46,24 +45,54 @@ class CVMUSBReadoutList;
 class C3820 : public CReadoutHardware
 {
 private:
+#ifdef MVLC_GENERATOR
+  XXUSB::CConfigurableObject*   m_pConfiguration;
+#else
   CReadoutModule*   m_pConfiguration;
+#endif
 public:
   C3820();
-  C3820(const C3820& rhs);
   virtual ~C3820();
+#ifdef MVLC_GENERATOR
+private:
+#endif
+  C3820(const C3820& rhs);
   C3820& operator=(const C3820& rhs);
+public:
 private:
   int operator==(const C3820& rhs) const;
   int operator!=(const C3820& rhs) const;
 
 public:
+#ifdef MVLC_GENERATOR
+  virtual void onAttach(XXUSB::CConfigurableObject& configuration);
+#else
   virtual void onAttach(CReadoutModule& configuration);
+#endif
   virtual void Initialize(CVMUSB& controller);
   virtual void addReadoutList(CVMUSBReadoutList& list);
+#ifndef MVLC_GENERATOR
   virtual CReadoutHardware* clone() const;
+#endif
 private:
   uint32_t getBase() const;
 };
 
+
+#ifdef MVLC_GENERATOR
+
+/**
+ * @class C3820Command
+ *     Derives from DeviceCommand to create ReadoutModules that wrap a C3820 device driver.
+ */
+class C3820Command : public DeviceCommand {
+public:
+  C3820Command(CTCLInterpreter& interp, TCLConfigParser& parser);
+  virtual ~C3820Command();
+protected:
+  virtual CReadoutModule* createDevice(std::string name);
+  
+};
+#endif
 
 #endif
